+++ src/overlords/defense/bunkerDefense.ts
@@ -3,7 +3,7 @@ import {CombatSetups, Roles} from '../../creepSetups/setups';
 import {DirectiveInvasionDefense} from '../../directives/defense/invasionDefense';
 import {OverlordPriority} from '../../priorities/priorities_overlords';
 import {profile} from '../../profiler/decorator';
-import {boostTypesAndTiers} from '../../resources/map_resources';
+import {BOOST_TIERS} from '../../resources/map_resources';
 import {CombatZerg} from '../../zerg/CombatZerg';
 import {CombatOverlord} from '../CombatOverlord';
 
@@ -25,7 +25,7 @@ export class BunkerDefenseOverlord extends CombatOverlord {
 		// Only spawn inside room
 		super(directive, 'bunkerDefense', priority, 1, 30);
 		this.lurkers = this.combatZerg(Roles.bunkerGuard, {
-			boostWishlist: boosted ? [boostTypesAndTiers.attack[3], boostTypesAndTiers.move[3]]
+			boostWishlist: boosted ? [BOOST_TIERS.attack.T3, BOOST_TIERS.move.T3]
 								   : undefined
 		});
 	}